feat: 상수 변수 util 파일로 이동
