+++ src/components/Manager/User/UserCopyModal.tsx
@@ -8,7 +8,11 @@ import {
 import { useEffect, useRef, useState } from "react";
 import { styled } from "styled-components";
 
-import { handleUserInfoForOrder } from "@/components/Manager/User/util";
+import {
+  handleUserInfoForOrder,
+  initialSelectedOptionObj,
+  optionsToCopyList,
+} from "@/components/Manager/User/util";
 import { OrdererInfo } from "@/const";
 
 const StyledModalContainer = styled.div`
@@ -20,43 +24,6 @@ const StyledModalContainer = styled.div`
   width: 400px;
   background-color: var(--color-white);
 `;
-
-interface OptionsToCopy {
-  label: string;
-  value: string;
-}
-
-const initialSelectedOptionObj = {
-  userId: true,
-  name: true,
-  companyName: true,
-  businessType: true,
-  phoneNumber: true,
-  weChatId: true,
-  email: true,
-  companyAddress: true,
-  shippingAddress: true,
-  hopeProducts: true,
-  productLengthUnit: true,
-  remark1: true,
-  remark2: true,
-};
-
-const optionsToCopyList: OptionsToCopy[] = [
-  { label: "USER ID", value: "userId" },
-  { label: "NAME", value: "name" },
-  { label: "COMPANY NAME", value: "companyName" },
-  { label: "BUSINESS TYPE", value: "businessType" },
-  { label: "PHONE NUMBER", value: "phoneNumber" },
-  { label: "WECHAT ID", value: "weChatId" },
-  { label: "EMAIL", value: "email" },
-  { label: "COMPANY ADDRESS", value: "companyAddress" },
-  { label: "SHIPPING ADDRESS", value: "shippingAddress" },
-  { label: "SELECTED ARTICLE LIST", value: "hopeProducts" },
-  { label: "LENGTH UNIT", value: "productLengthUnit" },
-  { label: "REMARK 1", value: "remark1" },
-  { label: "REMARK 2", value: "remark2" },
-];
 
 const UserCopyModal = ({
   isModalOpen,
