+++ src/components/Manager/User/util.ts
@@ -78,3 +78,40 @@ export const handleUserInfoForOrder = (
     })
     .flat();
 };
+
+export const initialSelectedOptionObj = {
+  userId: true,
+  name: true,
+  companyName: true,
+  businessType: true,
+  phoneNumber: true,
+  weChatId: true,
+  email: true,
+  companyAddress: true,
+  shippingAddress: true,
+  hopeProducts: true,
+  productLengthUnit: true,
+  remark1: true,
+  remark2: true,
+};
+
+interface OptionsToCopy {
+  label: string;
+  value: string;
+}
+
+export const optionsToCopyList: OptionsToCopy[] = [
+  { label: "USER ID", value: "userId" },
+  { label: "NAME", value: "name" },
+  { label: "COMPANY NAME", value: "companyName" },
+  { label: "BUSINESS TYPE", value: "businessType" },
+  { label: "PHONE NUMBER", value: "phoneNumber" },
+  { label: "WECHAT ID", value: "weChatId" },
+  { label: "EMAIL", value: "email" },
+  { label: "COMPANY ADDRESS", value: "companyAddress" },
+  { label: "SHIPPING ADDRESS", value: "shippingAddress" },
+  { label: "SELECTED ARTICLE LIST", value: "hopeProducts" },
+  { label: "LENGTH UNIT", value: "productLengthUnit" },
+  { label: "REMARK 1", value: "remark1" },
+  { label: "REMARK 2", value: "remark2" },
+];
